+++ Editor/d4rkAvatarOptimizerEditor.cs
@@ -610,11 +610,23 @@ public class d4rkAvatarOptimizerEditor : Editor
         if (hasVRCFuryComponent)
         {
             EditorGUILayout.HelpBox(
-                "VRCFury is on the avatar. This means the perf rank change and merge result previews can be inaccurate as the optimizer does not take VRCFury into account for those.\n" +
+                "VRCFury is used on the avatar. This means the perf rank change and merge result previews can be inaccurate as the optimizer does not take VRCFury into account for those.\n" +
                 "To test in editor built a VRCFury test avatar and use the optimizer on that.\n" +
-                "For uploading just use the Optimize on Upload feature as that ensures fury and the optimizer get used in the correct order.", MessageType.Warning);
+                "For uploading use the Optimize on Upload feature as that ensures fury and the optimizer get used in the correct order.", MessageType.Warning);
             return false;
         }
+
+        #if MODULAR_AVATAR_EXISTS
+        if (optimizer.GetComponentsInChildren<nadena.dev.modular_avatar.core.AvatarTagComponent>(true).Any())
+        {
+            EditorGUILayout.HelpBox(
+                "Modular Avatar is used on the avatar. This means the perf rank change and merge result previews " + 
+                "can be inaccurate as the optimizer does not take Modular Avatar into account for those.\n" +
+                "To test in editor use \"Manual bake avatar\" before clicking the optimize button.\n" +
+                "For uploading use the Optimize on Upload feature as that ensures Modular Avatar and the optimizer get used in the correct order.", MessageType.Warning);
+            return false;
+        }
+        #endif
 
         return true;
     }