+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add option to disable phys bones together with the skinned mesh they affect.
 * "Optimize FX Layer" can now also merge toggle layers that use two bools.
 * Add support for `.orlshader` files from [orels-Unity-Shaders](https://github.com/orels1/orels-Unity-Shaders/tree/main)
+* Add basic support for Modular Avatar by changing the callback order to -15 when Modular Avatar is detected.
 
 ### Changes
 * Renamed "Merge Simple Toggles as BlendTree" to "Optimize FX Layer".